pyproject.toml
[project]
name = "dddguardrails"
version = "0.1.0"
description = "3D guardrails"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.123.2",
    "python-multipart>=0.0.9",
    "trimesh>=4.4.3",
    "pyrender>=0.1.45",
    "numpy>=2.1.3",
    "Pillow>=11.0.0",
    "openai>=1.51.0",
    "pyglet<2",
    "pydantic-settings>=2.12.0",
    "pytest>=8.3.3",
    "google-genai>=1.53.0",
]


[build-system]
requires = ["uv_build>=0.9.14,<0.10.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
markers = [
    "e2e: end-to-end tests that call the real API",
]
